requirements.txt
torch
torchvision
transformers
Pillow
numpy
matplotlib
accelerate